Agregado y documentado, conexion con el AOuth del Front [Todo Funcional]
